# Zero-Copy Scans in the Vectorized Backend