--- a/src/exec/runners/PipelineRunner.h
+++ b/src/exec/runners/PipelineRunner.h
@@ -26,7 +26,7 @@ struct PipelineRunner {
    /// Run a single morsel of the backing pipeline.
    /// @param force_pick should we always pick, even if we are not a fuse chunk source?
    /// @return result of picking a morsel.
-   Suboperator::PickMorselResult runMorsel(size_t thread_id, bool force_pick);
+   virtual Suboperator::PickMorselResult runMorsel(size_t thread_id, bool force_pick);
 
    /// Clean up the intermediate morsel state from a previous failure.
    /// Purges the morsel size of the sinks to make sure we get a fresh