One big disadvantage our vectorized backend had so far was that it could not perform zero-copy table scans. Specifically, our compiling backend scans by directly taking raw pointers from the storage engine and feeding tuples into the computational pipeline through that.

Our vectorized backend meanwhile had to always copy the raw source data into an initial set of FuseChunks. This was an artifact of just interpreting the suboperators one-by-one.

The table scan IU providers would also be interpreted, and perform the initial storage -> fuse chunk copy. Successive vectorized primitives would then always work on top of these initial FuseChunks.

This commit changes the behaviour of the vectorized backend to also be able to scan directly from the storage engine.
This is implemented by changing the `runMorsel` implementation in the `InterpretedRunner` when seeing that we interpret a `TableScanIUProvider`.

In these cases we don't call into the pre-compiled primitive anymore, but rather just rewire pointers from within the FuseChunk sink column to directly point to the picked morsel in the storage engine.

The performance benefits of this approach are substantial. At SF5, we see the following improvements on my development machine:
```
Q1:  111 ms ->  82 ms
Q3:   76 ms ->  71 ms
Q4:   47 ms ->  43 ms
Q6:   31 ms ->  24 ms
Q14:  29 ms ->  16 ms
Q18:  600ms -> 593 ms
```
Unsurprisingly, the benefits are highest in scan-heavy queries.